Version 1.6.5
git-svn-id: https://plugins.svn.wordpress.org/recent-tweets-widget/trunk@1362258 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ settings.php
@@ -2,7 +2,7 @@
 add_thickbox();
 $tp_twitter_plugin_options = get_option('tp_twitter_plugin_options');
 
-if ($_GET['tp_twitter_global_notification'] == 0) {
+if (array_key_exists('tp_twitter_global_notification', $_GET) && $_GET['tp_twitter_global_notification'] == 0) {
         update_option('tp_twitter_global_notification', 0);
 }
 ?>
